move snapd enums back to snapdapi now they are no longer part of the subiquity API
